Update socket error in issue, kanban, milestone (Close #28)

+++ server/src/module/data/issueManager.ts
@@ -96,7 +96,7 @@ export default class DataIssueManager {
             fs.mkdirSync(this.getIssueInfoPath(kanbanUUID, issueUUID), { recursive: true });
         }
 
-        const issueData = { ...issueListData, content, milestone } as TIssueData;
+        const issueData = { ...issueListData, content, milestone, kanban: kanbanUUID } as TIssueData;
         if (!this.setIssueInfo(kanbanUUID, issueUUID, issueData)) {
             log.error(`[dataIssueManager] create -> fail to setIssueInfo`);
             return undefined;
@@ -117,9 +117,18 @@ export default class DataIssueManager {
             return false;
         }
         const beforeColumn = (issueListJsonData[uuid] as TIssueListData).column;
-        const issueListData = { ...issueListJsonData[uuid], uuid, issueId, title, creator, assigner, label, column } as TIssueListData;
 
-        if (!this.setIssueListInfo(kanbanUUID, uuid, issueListData)) {
+        if (
+            !this.setIssueListInfo(kanbanUUID, uuid, {
+                uuid: uuid,
+                issueId: issueId ?? issueListJsonData[uuid].issueId,
+                title: title ?? issueListJsonData[uuid].title,
+                creator: creator ?? issueListJsonData[uuid].creator,
+                assigner: assigner ?? issueListJsonData[uuid].assigner,
+                label: label ?? issueListJsonData[uuid].label,
+                column: column ?? issueListJsonData[uuid].column,
+            })
+        ) {
             log.error(`[dataIssueManager] update -> fail to setIssueListInfo`);
             return false;
         }
@@ -128,7 +137,20 @@ export default class DataIssueManager {
             log.error(`[dataIssueManager] update -> issueData is undefined`);
             return false;
         }
-        if (!this.setIssueInfo(kanbanUUID, uuid, { ...issueData, title, creator, assigner, label, column, content, milestone } as TIssueData)) {
+        if (
+            !this.setIssueInfo(kanbanUUID, uuid, {
+                uuid: uuid,
+                issueId: issueData.issueId,
+                title: title ?? issueData.title,
+                creator: creator ?? issueData.creator,
+                assigner: assigner ?? issueData.assigner,
+                label: label ?? issueData.label,
+                column: column ?? issueData.column,
+                content: content ?? issueData.content,
+                milestone: milestone ?? issueData.milestone,
+                kanban: issueData.kanban,
+            } as TIssueData)
+        ) {
             log.error(`[dataIssueManager] update -> fail to setIssueInfo`);
             return false;
         }
@@ -138,18 +160,27 @@ export default class DataIssueManager {
         if (beforeColumn !== "Done" && column === "Done") {
             DataKanbanManager.updateIssueCount(kanbanUUID, "doneIssue", "increase");
         }
-        log.info(`issue updated: ${issueData}`);
+        log.info(`issue updated: ${JSON.stringify(issueData)}`);
         return true;
     }
 
     static delete(kanbanUUID: string, issueUUID: string) {
         const issueListJsonData = this.getIssueListInfo(kanbanUUID);
         if (issueListJsonData === undefined) {
+            log.error(`[dataIssueManager] delete -> isueListJsonData is undefined`);
             return false;
         }
+        if (issueListJsonData[issueUUID].column === "Done") {
+            DataKanbanManager.updateIssueCount(kanbanUUID, "doneIssue", "decrease");
+        }
         delete issueListJsonData[issueUUID];
+        if (!setJsonData(this.getIssueListPath(kanbanUUID, "issueList.json"), issueListJsonData)) {
+            log.error(`[dataIssueManager] delete -> fail to delete issueData from issueList.json`);
+            return false;
+        }
         fs.rmdirSync(this.getIssueInfoPath(kanbanUUID, issueUUID), { recursive: true });
         DataKanbanManager.updateIssueCount(kanbanUUID, "totalIssue", "decrease");
+        log.info(`issue deleted: ${issueUUID}`);
         return true;
     }
 }
